requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=pyhd3eb1b0_0
arrow=1.2.3=py311haa95532_1
astroid=2.14.2=py311haa95532_0
asttokens=2.0.5=pyhd3eb1b0_0
atomicwrites=1.4.0=py_0
attrs=22.1.0=py311haa95532_0
autopep8=1.6.0=pyhd3eb1b0_1
babel=2.11.0=py311haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py311h2bbff1b_1
beautifulsoup4=4.12.2=py311haa95532_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=23.3.0=py311haa95532_0
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py311h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=py311haa95532_0
cffi=1.15.1=py311h2bbff1b_3
chardet=4.0.0=py311haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311haa95532_0
cloudpickle=2.2.1=py311haa95532_0
colorama=0.4.6=py311haa95532_0
comm=0.1.2=py311haa95532_0
cookiecutter=1.7.3=pyhd3eb1b0_0
cryptography=41.0.2=py311hac1b9e3_0
debugpy=1.6.7=py311hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=pyhd3eb1b0_0
dill=0.3.7=py311haa95532_0
docstring-to-markdown=0.11=py311haa95532_0
docutils=0.18.1=py311haa95532_3
entrypoints=0.4=py311haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flake8=6.0.0=py311haa95532_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
icu=58.2=ha925a31_3
idna=3.4=py311haa95532_0
imagesize=1.4.1=py311haa95532_0
importlib-metadata=6.0.0=py311haa95532_0
importlib_metadata=6.0.0=hd3eb1b0_0
inflection=0.5.1=py311haa95532_0
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.25.0=py311h746a85d_0
ipython=8.12.2=py311haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.9.3=pyhd3eb1b0_0
jaraco.classes=3.2.1=pyhd3eb1b0_0
jedi=0.18.1=py311haa95532_1
jellyfish=0.9.0=py311h2bbff1b_0
jinja2=3.1.2=py311haa95532_0
jinja2-time=0.2.0=pyhd3eb1b0_3
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=py311haa95532_0
jupyter_client=8.1.0=py311haa95532_0
jupyter_core=5.3.0=py311haa95532_0
jupyterlab_pygments=0.1.2=py_0
keyring=23.13.1=py311haa95532_0
krb5=1.20.1=h5b6d351_0
lazy-object-proxy=1.6.0=py311h2bbff1b_0
lerc=3.0=hd77b12b_0
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.5.1=hd77b12b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lxml=4.9.2=py311h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py311h2bbff1b_0
matplotlib-inline=0.1.6=py311haa95532_0
mccabe=0.7.0=pyhd3eb1b0_0
mistune=0.8.4=py311h2bbff1b_1000
more-itertools=8.12.0=pyhd3eb1b0_0
mypy_extensions=0.4.3=py311haa95532_1
nbclient=0.5.13=py311haa95532_0
nbconvert=6.5.4=py311haa95532_0
nbformat=5.7.0=py311haa95532_0
nest-asyncio=1.5.6=py311haa95532_0
numpydoc=1.5.0=py311haa95532_0
openssl=3.0.10=h2bbff1b_2
packaging=23.1=py311haa95532_0
pandocfilters=1.5.0=pyhd3eb1b0_0
paramiko=2.8.1=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.10.3=py311haa95532_0
pcre=8.45=hd77b12b_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=23.2.1=py311haa95532_0
platformdirs=3.10.0=py311haa95532_0
pluggy=1.0.0=py311haa95532_1
ply=3.11=py311haa95532_0
poyo=0.5.0=pyhd3eb1b0_0
prompt-toolkit=3.0.36=py311haa95532_0
psutil=5.9.0=py311h2bbff1b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycodestyle=2.10.0=py311haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pydocstyle=6.3.0=py311haa95532_0
pyflakes=3.0.1=py311haa95532_0
pygments=2.15.1=py311haa95532_1
pylint=2.16.2=py311haa95532_0
pylint-venv=2.3.0=py311haa95532_0
pyls-spyder=0.4.0=pyhd3eb1b0_0
pynacl=1.5.0=py311h8cc25b3_0
pyopenssl=23.2.0=py311haa95532_0
pyqt=5.15.7=py311hd77b12b_0
pyqt5-sip=12.11.0=py311hd77b12b_0
pyqtwebengine=5.15.7=py311hd77b12b_0
pyrsistent=0.18.0=py311h2bbff1b_0
pysocks=1.7.1=py311haa95532_0
python=3.11.4=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311haa95532_0
python-lsp-black=1.2.1=py311haa95532_0
python-lsp-jsonrpc=1.0.0=pyhd3eb1b0_0
python-lsp-server=1.7.2=py311haa95532_0
python-slugify=5.0.2=pyhd3eb1b0_0
pytoolconfig=1.2.5=py311haa95532_1
pytz=2022.7=py311haa95532_0
pywin32=305=py311h2bbff1b_0
pywin32-ctypes=0.2.0=py311haa95532_1000
pyyaml=6.0=py311h2bbff1b_1
pyzmq=25.1.0=py311hd77b12b_0
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.2.2=py311haa95532_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtawesome=1.2.2=py311haa95532_0
qtconsole=5.4.2=py311haa95532_0
qtpy=2.2.0=py311haa95532_0
qtwebkit=5.212=h2bbfb41_5
requests=2.31.0=py311haa95532_0
rope=1.7.0=py311haa95532_0
rtree=1.0.1=py311h2eaa2aa_0
setuptools=68.0.0=py311haa95532_0
sip=6.6.2=py311hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snowballstemmer=2.2.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.4=py311haa95532_0
sphinx=5.0.2=py311haa95532_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=5.4.3=py311haa95532_1
spyder-kernels=2.4.4=py311haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
text-unidecode=1.3=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tinycss2=1.2.1=py311haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomlkit=0.11.1=py311haa95532_0
tornado=6.3.2=py311h2bbff1b_0
traitlets=5.7.1=py311haa95532_0
tzdata=2023c=h04d1e81_0
ujson=5.4.0=py311hd77b12b_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.16=py311haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py311haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311haa95532_1
whatthepatch=1.0.2=py311haa95532_0
wheel=0.38.4=py311haa95532_0
win_inet_pton=1.1.0=py311haa95532_0
wrapt=1.14.1=py311h2bbff1b_0
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py311haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
